SSRCHECKER/aom/initialization_centOS7_forwarding_with_puppet_2018_variable.py
#!/usr/bin/python3
import yaml
filename = "initialization-centOS7-forwarding-with-puppet-2018.yml"
with open(filename) as f:
    fyaml = yaml.load(f, Loader=yaml.FullLoader)
test = fyaml[0]
print(test)

