requirements.txt
amqp==2.2.2
asn1crypto==0.22.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.0
billiard==3.5.0.3
bs4==0.0.1
cachetools==2.0.1
celery==4.2.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
crypto==1.4.1
defusedxml==0.5.0
Django==1.11.3
django-braces==1.12.0
django-easy-select2==1.3.4
django-filter==1.0.4
dnspython==1.15.0
docker-py==1.10.6
docker-pycreds==0.3.0
EasyProcess==0.2.3
eight==0.4.2
enum34==1.1.6
fake-useragent==0.1.7
future==0.16.0
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.9.0
idna==2.5
ipaddress==1.0.18
kombu==4.2.1
lxml==3.8.0
Markdown==2.6.8
Naked==0.1.31
numpy==1.15.1
pycparser==2.18
pycrypto==2.6.1
pyOpenSSL==17.2.0
PySocks==1.6.8
python-nginx==1.4.1
python-whois==0.6.5
pytz==2017.2
PyVirtualDisplay==0.2.1
pywhois==0.1
PyYAML==3.12
redis==2.10.6
requests==2.18.3
requests-file==1.4.2
selenium==3.14.0
shellescape==3.4.1
signxml==2.4.0
six==1.10.0
tldextract==2.1.0
urllib3==1.22
vine==1.1.4
websocket==0.2.1
websocket-client==0.44.0
xvfbwrapper==0.2.9
whois==0.9.7
tld==0.12.2
